+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file (focus on ch
 
 - Fixes:
 	- wifi-first-activation-not-list-proposed #[13]
+- Features:
+    - Add volume only for video boot
 
 ## [recalbox-integration] - 2022-09-12 - v0.0.8